b1c1: Decrease reserved space on /system
Some apps in LineageOS have grown significantly in recent weeks.

To name a few:
webview: 155MB to 180MB
Seedvault: 11.5MB to 22.5MB
LineageSetupWizard: 2.5MB to 7.5MB

Which caused the build to fail due to not enough free space available.
This change gives it 50MB of free space for future growth.

Change-Id: Ib77f748c2cad6785e99f2b9ec809c71152d0b2db